pyproject.toml
[project]
name = "spotter-ai-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django-environ>=0.12.0",
    "django>=5.1.7",
    "django-cors-headers>=4.7.0",
    "djangorestframework>=3.15.2",
    "djantic>=0.7.0",
    "gunicorn>=23.0.0",
    "pillow>=11.1.0",
    "pydantic>=2.10.6",
    "pyngo>=2.3.0",
    "whitenoise>=6.9.0",
]
